infra: rename co_spawn_sw header as spawn
