+++ silkworm/infra/concurrency/awaitable_wait_for_all.hpp
@@ -33,8 +33,8 @@
 #include <boost/asio/this_coro.hpp>
 #include <boost/asio/use_awaitable.hpp>
 
-#include "co_spawn_sw.hpp"
 #include "parallel_group_utils.hpp"
+#include "spawn.hpp"
 
 namespace silkworm::concurrency::awaitable_wait_for_all {
