+++ silkworm/node/execution/api/active_direct_service.cpp
@@ -16,7 +16,7 @@
 
 #include "active_direct_service.hpp"
 
-#include <silkworm/infra/concurrency/co_spawn_sw.hpp>
+#include <silkworm/infra/concurrency/spawn.hpp>
 
 namespace silkworm::execution::api {
