+++ silkworm/node/stagedsync/stages/stage_triggers.cpp
@@ -20,7 +20,7 @@
 
 #include <gsl/util>
 
-#include <silkworm/infra/concurrency/co_spawn_sw.hpp>
+#include <silkworm/infra/concurrency/spawn.hpp>
 
 namespace silkworm::stagedsync {
